requirements.txt
astroid==1.6.5
Django==2.1.2
django-appconf==1.0.2
django-imagekit==4.0.2
ikp3db==1.1.4
isort==4.3.4
jedi==0.11.1
lazy-object-proxy==1.3.1
mccabe==0.6.1
parso==0.1.1
pilkit==2.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.4
pytz==2018.7
six==1.11.0
virtualenv==16.1.0
wrapt==1.10.11